requirements.txt
autopep8
Django
django-cors-headers
django-recurrence
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
PyJWT
pylint
pylint-django
pylint-plugin-utils
Pillow